Reorging plotting into plotting.py, better scenario control.
